Fix : Reduce uint, str & log complexity

+++ toolkit/src/training/ovtkCSignalTrialSet.hpp
@@ -12,7 +12,7 @@ namespace OpenViBEToolkit
 		bool addSignalTrial(ISignalTrial& rSignalTrial) override;
 		bool clear() override;
 		uint32_t getSignalTrialCount() const override { return uint32_t(m_vSignalTrial.size()); }
-		ISignalTrial& getSignalTrial(const uint32_t index) const override { return *m_vSignalTrial[index]; }
+		ISignalTrial& getSignalTrial(const size_t index) const override { return *m_vSignalTrial[index]; }
 
 		_IsDerivedFromClass_Final_(OpenViBEToolkit::ISignalTrialSet, OVTK_ClassId_)
